# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Store API",
        "description": "This is the API for managing stores.",
        "version": "1.0.0"
    },
    "paths": {
        "/api/stores": {
            "get": {
                "tags": [
                    "Store"
                ],
                "summary": "Get stores near a given postcode",
                "description": "Retrieve a list of stores near a given postcode with optional delivery information.",
                "operationId": "getStoresNearby",
                "parameters": [
                    {
                        "name": "postcode",
                        "in": "query",
                        "description": "The postcode of the location to find nearby stores",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "delivery",
                        "in": "query",
                        "description": "Whether to filter stores that offer delivery (true/false)",
                        "required": false,
                        "schema": {
                            "type": "boolean"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Stores retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Stores retrieved successfully."
                                        },
                                        "stores": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "description": "Store ID",
                                                        "type": "integer"
                                                    },
                                                    "name": {
                                                        "description": "Store name",
                                                        "type": "string"
                                                    },
                                                    "location": {
                                                        "properties": {
                                                            "type": {
                                                                "type": "string",
                                                                "example": "Point"
                                                            },
                                                            "coordinates": {
                                                                "type": "array",
                                                                "items": {
                                                                    "type": "number",
                                                                    "format": "float",
                                                                    "example": -4.006499
                                                                }
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "status": {
                                                        "description": "Store status (e.g., open/closed)",
                                                        "type": "string"
                                                    },
                                                    "type": {
                                                        "description": "Store type (e.g., shop)",
                                                        "type": "string"
                                                    },
                                                    "max_delivery_distance": {
                                                        "description": "Maximum delivery distance in meters",
                                                        "type": "integer"
                                                    },
                                                    "created_at": {
                                                        "description": "Store creation timestamp",
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "updated_at": {
                                                        "description": "Store update timestamp",
                                                        "type": "string",
                                                        "format": "date-time"
                                                    },
                                                    "distance": {
                                                        "description": "Distance from the given location in meters",
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "next_cursor": {
                                                    "description": "Next page cursor, if available",
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "previous_cursor": {
                                                    "description": "Previous page cursor, if available",
                                                    "type": "string",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid postcode or query parameters",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid postcode or delivery filter."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No stores found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No stores found near the given location"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Store"
                ],
                "summary": "Create a new store",
                "operationId": "0819446609f0655cb170f575d9ac2ddc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "location",
                                    "status",
                                    "type",
                                    "max_delivery_distance"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Name of the store",
                                        "type": "string"
                                    },
                                    "location": {
                                        "description": "Geographical location of the store [longitude, latitude]",
                                        "type": "array",
                                        "items": {
                                            "type": "number"
                                        }
                                    },
                                    "status": {
                                        "description": "Status of the store",
                                        "type": "string"
                                    },
                                    "type": {
                                        "description": "Type of the store",
                                        "type": "string"
                                    },
                                    "max_delivery_distance": {
                                        "description": "Maximum delivery distance in meters",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Store created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Store created successfully."
                                        },
                                        "store": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": "1"
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Store Name"
                                                },
                                                "location": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "number"
                                                    },
                                                    "example": [
                                                        -0.1276,
                                                        51.5074
                                                    ]
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "open"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "grocery"
                                                },
                                                "max_delivery_distance": {
                                                    "type": "integer",
                                                    "example": 5000
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Invalid input"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Store",
            "description": "Store"
        }
    ]
}